refactor: refactor back buttons sign up flow
