+++ src/routes/index.js
@@ -1,9 +1,9 @@
-export * from "./account";
+// export * from "./account";
 export * from "./account-page";
 export * from "./add-account";
 export * from "./home";
-export * from "./login";
 export * from "./profile";
 export * from "./sign-up";
 export * from "./steps";
 export * from "./temp-stub";
+export * from "./unused-temp-storage/login";
