+++ src/routes/steps/seed-phrase/seed-phrase.jsx
@@ -41,6 +41,8 @@ export function SeedPhrase({
               buttonKind="textOnly"
               leftIcon="arrowLeft"
               size="sm"
+              margin="0"
+              className="px-0"
               onClick={onBackClick}
             />
           </Layout.Header.Left>
