+++ src/routes/steps/user-info/user-info.jsx
@@ -27,6 +27,8 @@ export function UserInfo({ onBackClick, onForwardClick, title, subtitle }) {
               buttonKind="textOnly"
               leftIcon="arrowLeft"
               size="sm"
+              margin="0"
+              className="px-0"
               onClick={onBackClick}
             />
           </Layout.Header.Left>
