+++ src/routes/steps/wallet-confirm/wallet-confirm.jsx
@@ -57,6 +57,7 @@ export const WalletConfirm = ({
               buttonKind="textOnly"
               leftIcon="arrowLeft"
               size="sm"
+              className="p-0 m-0"
               onClick={onBackClick}
             />
           </Layout.Header.Left>
