+++ src/routes/steps/wallet-scan/wallet-scan.jsx
@@ -5,14 +5,12 @@ import { account, fSeedPhrases } from "../../../utils/faker-data";
 import { Body } from "./body";
 
 //TODO: implement qr scanner component
-//TODO:logic for grabbing qr values with faker- tbd with SDK-
+//TODO:logic for grabbing qr values for SDK-
 //consider clearing state.accountData before scanning
 //TODO: logic to get seeds from scan and pass to context - working with faker
 //TODO: what should happen on public key manually
 //a form open or turn into input field
 //how will you know if submitted without button?  will continue serve as submit?
-//TODO: get form inputs when ready for input public key
-//may need styled ocmponents to render placeholder text like figma
 
 /**
  * Wallet-Scan Component -
@@ -53,7 +51,7 @@ export const WalletScan = ({
   const [data, setData] = useState("");
   const [key, setKey] = useState("");
 
-  //state of scan data capture
+  //state of scan data capture - to be hooked up with scanner
   const [didGetData, setDidGetData] = useState(false);
 
   //control buuton state
@@ -77,12 +75,12 @@ export const WalletScan = ({
     }
   }, [state.accountData, key]);
   // /* --------------------------------- handler -------------------------------- */
-  //w react-qrreader - passed via onResult?how to implement
+
   //currently pulling account data and seed phrases(needed for next step) from faker
   const handleScan = () => {
     setData(account);
     updateSeedPhrases(fSeedPhrases);
-    console.log(fSeedPhrases);
+
     //sdk will handle this
   };
